refactor pq_kmeans to file
