+++ src/algs/product_quantization.rs
@@ -1,33 +1,15 @@
-use ndarray::{Array, Array1, Array2, Array3, ArrayView1, ArrayView2, s};
-// use crate::algs::data_entry::{DataEntry};
+use ndarray::{Array, Array1, Array2, ArrayView1, ArrayView2, s};
 use crate::algs::*;
-// use std::collections::BinaryHeap;
-use rand::prelude::*;
 use rand::{distributions::Uniform, Rng};
 use std::collections::HashMap;
 use colored::*;
-
-#[derive(Clone, PartialEq, Debug)]
-pub struct Centroid {
-    pub point: Array1::<f64>,
-    pub children: Vec::<usize>
-}
-
-impl Centroid {
-    fn new(id: i32, point: Array1::<f64>) -> Self {
-        Centroid {
-            point: point,
-            children: Vec::<usize>::new()
-        }
-    }
-}
+use pq_kmeans::{PQKMeans};
 
 #[derive(Debug, Clone)]
 pub struct ProductQuantization {
     name: String,
     metric: String,
     dataset: Option<Array2::<f64>>,
-    codebook: Vec<Vec<Centroid>>,
     pqcodes: Option::<Array2::<usize>>,
     m: usize,
     training_size: usize,
@@ -46,7 +28,6 @@ impl ProductQuantization {
             name: "FANN_product_quantization()".to_string(),
             metric: "angular".to_string(),
             dataset: None,
-            codebook: Vec::with_capacity(m),
             pqcodes: None,
             m: m,         // M
             training_size: training_size,
@@ -56,22 +37,6 @@ impl ProductQuantization {
             verbose_print: verbose_print,
             dimension: 0,
             sub_dimension: 0
-        }
-    }
-
-    fn _print_sum_codebook_children(&self, info: &str, codebook: &HashMap<i32, Centroid>, dataset_len: usize) {
-        println!("{}", info.to_string().on_white().black());
-        let mut sum = 0;
-        for (_, centroid) in codebook.iter() {
-            sum += centroid.children.len();
-        }
-        println!("children: {} == {} dataset points, equal {}", sum.to_string().blue(), dataset_len.to_string().blue(), (sum == dataset_len).to_string().blue());
-    }
-    
-    fn _print_codebook(&self, info: &str, codebook: &HashMap<i32, Centroid>) {
-        println!("{}", info.to_string().on_white().black());
-        for (key, centroid) in codebook.iter() {
-            println!("-> centroid C{:?} |  children: {:?} | point sum: {:?}", key, centroid.children.len(), centroid.point.sum());
         }
     }
 
@@ -106,8 +71,8 @@ impl ProductQuantization {
             if self.verbose_print {
                 println!("Run k-means for m [{}], sub dim {:?}, first element [{}]", m, partial_data.shape(), partial_data[[0,0]]);
             }
-            let mut kmeans = KMeans::new(self.k, self.max_iterations);
-            let codewords = kmeans.run(partial_data.view());
+            let mut pq_kmeans = PQKMeans::new(self.k, self.max_iterations);
+            let codewords = pq_kmeans.run(partial_data.view());
             for (k, (centroid,_)) in codewords.iter().enumerate() {
                     codebook[[m,k]] = centroid.to_owned();
             }
@@ -145,124 +110,6 @@ impl ProductQuantization {
 }
 
 
-#[derive(Debug, Clone)]
-struct KMeans {
-    k: usize,
-    max_iterations: usize,
-    codebook: Vec::<(Array1::<f64>, Vec::<usize>)>,
-    verbose_print: bool
-}
-
-impl KMeans {
-
-    pub fn new(k: usize, max_iterations: usize) -> Self {
-        KMeans{
-            k: k,
-            max_iterations: max_iterations,
-            codebook: Vec::with_capacity(k),
-            verbose_print: true
-        }
-    }
-
-    fn _print_codebook(&self, info: &str, codebook: &Vec::<(Array1::<f64>, Vec::<usize>)>) {
-        println!("{}", info.to_string().on_white().black());
-        for (k, (centroid, children)) in codebook.iter().enumerate() {
-            println!("-> centroid C{:?} |  children: {:?} | point sum: {:?}", k, children.len(), centroid.sum());
-        }
-    }
-
-    
-
-    pub fn run(&mut self, dataset: ArrayView2::<f64> ) -> &Vec::<(Array1::<f64>, Vec::<usize>)> {
-
-        self.codebook = Vec::with_capacity(self.k);
-        self.init(dataset);        
-        let mut last_codebook = Vec::with_capacity(self.k);
-        let mut iterations = 1;
-        loop {
-            if self.verbose_print && (iterations == 1 || iterations % 10 == 0) {
-                println!("Iteration {}", iterations);
-            }
-            if iterations > self.max_iterations {
-                if self.verbose_print {
-                    println!("Max iterations reached, iterations: {}", iterations-1);
-                }
-                break;
-            } else if self.codebook == last_codebook {
-                if self.verbose_print {
-                    println!("Computation has converged, iterations: {}", iterations-1);
-                }
-                break;
-            }
-    
-            last_codebook = self.codebook.clone();
-
-            self.assign(dataset);
-            self.update(dataset);
-            iterations += 1;
-        }
-        self._print_codebook("Codebook after run: ", &self.codebook);
-        return &self.codebook;
-    }
-
-    fn init(&mut self, dataset: ArrayView2::<f64>) {
-        let mut rng = thread_rng();
-        let dist_uniform = rand::distributions::Uniform::new_inclusive(0, dataset.nrows()-1);
-        for i in 0..self.k {
-            let rand_key = rng.sample(dist_uniform);
-            let candidate = dataset.slice(s![rand_key,..]);
-            self.codebook.push((candidate.to_owned(), Vec::<usize>::new()));
-        }
-
-        if self.verbose_print {
-            self._print_codebook("Codebook after init: ", &self.codebook);
-        }      
-    }
-
-    fn assign(&mut self, dataset: ArrayView2::<f64>) {
-        for (_,children) in self.codebook.iter_mut() {
-            children.clear();
-        }
-
-        for (idx, candidate) in dataset.outer_iter().enumerate() {
-            let mut best_centroid = None;
-            let mut best_distance = f64::NEG_INFINITY;
-            for (k, centroid) in self.codebook.iter().enumerate() {
-                let distance = distance::cosine_similarity(&(centroid.0).view(), &candidate);
-                if best_distance < distance {
-                    best_centroid = Some(k);
-                    best_distance = distance;
-                }
-            }
-            if best_centroid.is_some() {
-                self.codebook[best_centroid.unwrap()].1.push(idx);
-            } 
-        }
-    }
-
-    fn update(&mut self, dataset: ArrayView2::<f64>) {
-        for (centroid, childeren) in self.codebook.iter_mut() {
-            if childeren.len() > 0 {
-                for i in 0..centroid.len() {
-                    centroid[i]= 0.;
-                }
-                
-                for child_key in childeren.iter() {
-                    let child_point = dataset.slice(s![*child_key,..]);
-                    for (i, x) in child_point.iter().enumerate() {
-                        centroid[i] += x;
-                    }
-                }
-    
-                for i in 0..centroid.len() {  
-                    centroid[i] = centroid[i]/childeren.len() as f64;
-                }
-            }
-        }
-    }
-
-
-}
 
 impl AlgorithmImpl for ProductQuantization {
